refactor(settings): drop the colour format hover tooltip, keep the click
The dialog already carries the whole explanation, so a hover summary on top of
it was just one more thing to read past. The icon is now a plain IconButton with
no tooltip at all — and no IconButton `tooltip:` either, which would have been
the same hover by another route.

The accessible name moves to the icon's semanticLabel, so a screen reader still
announces it without a popup appearing for pointer users. Two tests pin the
decision: no Tooltip in the widget, and the semanticLabel is still there — the
second exists because dropping a tooltip is the easy way to drop an element's
accessible name along with it.

The tooltip contrast tests go with it (there is no tooltip left to be unreadable),
but the reasoning stays in the widget doc comment: Material's tooltip surface is
inverted, so anyone re-adding one needs onInverseSurface, not the page's text
colour. Test file renamed to match what it now covers.
